backend/app/schemas/common.py
"""Shared response models."""

from typing import Generic, TypeVar

from pydantic import BaseModel


T = TypeVar("T")


class ApiResponse(BaseModel, Generic[T]):
    """Unified API response envelope."""

    success: bool = True
    message: str = "ok"
    data: T | None = None
