pyproject.toml
[project]
name = "nicefk"
version = "0.1.0"
description = "nicefk 发卡网，基于 FastAPI + React + MySQL + Redis"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "pymysql>=1.1,<2.0",
  "redis>=5.0,<6.0",
  "apscheduler>=3.10,<4.0",
  "httpx>=0.27,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "cryptography>=45.0,<46.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=5.0,<6.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]

[tool.setuptools.package-dir]
backend = "backend"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["backend/tests"]
